fixed remaining bugs from refactor
