+++ frontend/new-react-vite/src/components/CyberNav.css
@@ -1,0 +1,19 @@
+.top {
+    flex: 0 0 auto; 
+    background-color: black; /* Optional: Background for the top */
+    color: white;
+    padding: 0 3%; 
+}
+.top-elements {
+    display: flex; 
+    align-items: center; /* center vertically */
+    justify-content: space-between; 
+    gap: 20px; 
+}
+.top-navbar {
+    margin: 0;
+    padding: 0;
+    list-style-type: none;
+    display: flex; 
+    gap: 30px; 
+}
