+++ frontend/new-react-vite/src/components/Marketplace/Marketplace.css
@@ -5,34 +5,10 @@ body {
     height: 100vh; 
 }
 
-.top {
-    flex: 0 0 auto; 
-    background-color: black; /* Optional: Background for the top */
-    color: white;
-    padding: 0 3%; 
-}
-.top-elements {
-    display: flex; 
-    align-items: center; /* center vertically */
-    justify-content: space-between; 
-    gap: 20px; 
-}
-.top-navbar {
-    margin: 0;
-    padding: 0;
-    list-style-type: none;
-    display: flex; 
-    gap: 30px; 
-}
-
-
-
 .left-right-container {
     display: flex;
     flex: 1; 
 }
-
-
 
 .left {
     flex: 1;
